General improvements to the rest API tests.
- Added "is_request_allowed" to check the PMPro toolkit protocol settings to ensure that it only allows the methods based on toolkit settings.

The reason it's done inside the `handle_request` method is to be an extra guard rail, and show a better error message than "not at all". While it makes sense to also only register the routes when the "read_write" is enabled, this is purely a clearer indication of "why" this route isn't returning or processing data. Not registering the route based on settings returns a 404 error and can be a bit confusing to developers without knowing that you need to enable read_write options.

It _can_ go either way, this felt like a better user approach.